Struct comms::room::Room

pub struct Room<I, C> where I: Clone + Send + PartialEq + Eq + Hash + Debug + 'static,
        C: Sink + Stream + 'static { /* fields omitted */ }

Handles connection with multiple server clients.

It value-adds a lot of consistency by being a well-developed group of clients. In my applications it's being immensely helpful to factor out this logic to here.

Methods

impl<I, C> Room<I, C> where I: Clone + Send + PartialEq + Eq + Hash + Debug + 'static,
C: Sink + Stream + 'static

fn new(clients: Vec<Client<I, C>>) -> Room<I, C>

Construct a new Room from a list of Client.

N.B., construct an empty room with Room::default().

fn ids(&self) -> HashSet<I>

Get the IDs of all connected clients.

fn client(&self, id: &I) -> Option<&Client<I, C>>

fn insert(&mut self, client: Client<I, C>) -> bool

Insert a Client into this room and return whether it was newly added.

For now this accepts disconnected clients but this may change in future.

fn remove(&mut self, id: &I) -> Option<Client<I, C>>

Remove a client by ID.

fn contains(&self, id: &I) -> bool

Check is a client ID is present

fn broadcast_all(self, msg: C::SinkItem) -> Broadcast<I, C> where C::SinkItem: Clone

Broadcast a single message to all connected clients.

fn broadcast(self, msg: C::SinkItem, ids: HashSet<I>) -> Broadcast<I, C> where C::SinkItem: Clone

Broadcast a single message to particular connected clients.

fn transmit(self, msgs: HashMap<I, C::SinkItem>) -> Transmit<I, C>

Send particular messages to particular connected clients.

fn receive_all(self) -> Receive<I, C>

Try to receive a single message from all clients.

fn receive(self, ids: HashSet<I>) -> Receive<I, C>

Try to receive a single message from particular clients.

fn close(&mut self, ids: HashSet<I>)

Force disconnection of particular clients.

fn close_all(&mut self)

Disconnect all clients.

fn into_clients(self) -> HashMap<I, Client<I, C>>